+++ arm-helper/src/arm-helper.ts
@@ -1,8 +1,9 @@
-import { IIngredient, Logger, DeploymentContext, BakeVariable, TagGenerator } from '@azbake/core';
+import { IIngredient, Logger, DeploymentContext, BakeVariable, TagGenerator, IngredientManager, objToVariableMap } from '@azbake/core';
 import { ResourceManagementClient } from '@azure/arm-resources';
 import { Deployment, DeploymentProperties } from '@azure/arm-resources/esm/models';
 import { stringify } from 'querystring';
 import { AnyCnameRecord } from 'dns';
+import alertTemplate from "./alert.json"
 
 export class ARMHelper {
 
@@ -36,6 +37,7 @@ export class ARMHelper {
                 }
             }
 
+            logger.log(`resource group: ${resourceGroup}`);
             logger.debug('template:\n' + JSON.stringify(template, null, 3));
             logger.debug('input params:\n' + JSON.stringify(params, null, 3));
 
@@ -74,6 +76,86 @@ export class ARMHelper {
             logger.error('deployment failed: ' + error);
             throw error;
         }
+    }
+
+    public async DeployAlerts(targetDeploymentName: string, resourceGroup: string, alertTarget: string, stockAlerts: any, alertsOverrides: Map<string, BakeVariable>): Promise<void> {
+        let deploymentName:string = 'alerts-' + targetDeploymentName;
+        const logger = new Logger(this._ctx.Logger.getPre().concat(deploymentName), this._ctx.Environment.logLevel);
+        
+        let alertOverridesParams = await this.BakeParamsToARMParamsAsync(deploymentName, alertsOverrides);
+        let stockAlertsMap = objToVariableMap(stockAlerts);
+        let stockAlertsParams = await this.BakeParamsToARMParamsAsync(deploymentName, stockAlertsMap);
+
+        for (let stockAlert in stockAlerts) {
+            let stockAlertMap = objToVariableMap(stockAlerts[stockAlert]);
+            let stockAlertParamsARM = await this.BakeParamsToARMParamsAsync(deploymentName, stockAlertMap);
+            let alertOverrideParams: any  | undefined
+            alertOverrideParams = alertsOverrides.get(stockAlert)
+            
+            let mergedAlertParamsARM: any, alertOverrideParamsARM: any;
+
+            if (alertOverrideParams !== undefined) {
+                let alertOverrideParamsMap = objToVariableMap(await alertOverrideParams.valueAsync(this._ctx))
+                alertOverrideParamsARM = await this.BakeParamsToARMParamsAsync(deploymentName, alertOverrideParamsMap)
+
+                mergedAlertParamsARM = this.mergeDeep(stockAlertParamsARM, alertOverrideParamsARM)
+            }
+            else {
+                mergedAlertParamsARM = stockAlertParamsARM
+            }
+
+            await this.DeployAlert(deploymentName, resourceGroup, alertTarget, mergedAlertParamsARM)
+
+        }
+    }
+
+    public async DeployAlert(deploymentName: string, resourceGroup: string, alertTarget: string, params: any): Promise<void> {
+        let util = IngredientManager.getIngredientFunction("coreutils", this._ctx)
+        
+        params["source-rg"] = { "value": resourceGroup };
+        params["source-name"] = { "value": alertTarget };
+
+        let timeAggregation = params["timeAggregation"].value;
+        let metricName:string = params["metricName"].value;
+        metricName = metricName.replace(/[/%]/g,'_')  //replace / and % with _.  They are valid metric name character but not valid in alert names.
+        let alertType = params["alertType"].value;
+        let tempName = '-' + alertTarget + '-' + timeAggregation + '-' + metricName + '-' + alertType;
+        let alertName:string = util.create_resource_name("alert", tempName, true);                        
+        alertName = alertName.substr(0, 128)    //Azure limits alert names to 128 characters
+        const logger = new Logger(this._ctx.Logger.getPre().concat(deploymentName), this._ctx.Environment.logLevel);
+        logger.log(alertName);            
+        params["alertName"] = { "value": alertName };
+
+        if (params["actionGroups"])
+        {
+            if (params["actionGroups"].value) {
+                params["actionGroups"].value.forEach(async  (element:any) => {
+                    if (element.actionGroupShortName) 
+                    {
+                        let actionGroupShortName = element.actionGroupShortName;
+                        
+                        let subscriptionId = this._ctx.Environment.authentication.subscriptionId;
+                        let actionResourceGroup = await util.resource_group("actiongroups",false,null,true); 
+                        let actionGroup = util.create_resource_name("act", actionGroupShortName, false);
+
+                        let actionGroupId = "/subscriptions/" + subscriptionId + "/resourceGroups/" + actionResourceGroup + "/providers/Microsoft.Insights/actionGroups/" + actionGroup                        
+
+                        element.actionGroupId = actionGroupId; //Add the actionGroupId param
+                        delete element.actionGroupShortName; //Remove the actionGroupShortName param
+                    }
+                });
+            }
+            else {
+                delete params["actionGroups"]; //Remove the actionGroups section if there aren't any specified
+            }
+        }
+
+        // Don't fail the Bake recipe if there are problems deploying an alert.  Instead, log as a warning and continue.
+        try {
+            await this.DeployTemplate(deploymentName, alertTemplate, params, resourceGroup);
+        } catch (error) {
+            logger.warn('Unable to deploy alert.  Continuing deployments.');
+        }        
     }
 
     public async BakeParamsToARMParamsAsync(deploymentName: string, params: Map<string, BakeVariable>): Promise<any> {
@@ -122,5 +204,27 @@ export class ARMHelper {
 
         template.resources = resources;
         return template;
+    }
+
+    private isObject(item:any) {
+        return (item && typeof item === 'object' && !Array.isArray(item));
+    }
+  
+    private mergeDeep(target:any, ...sources:any): any {
+        if (!sources.length) return target;
+        const source = sources.shift();
+    
+        if (this.isObject(target) && this.isObject(source)) {
+            for (const key in source) {
+                if (this.isObject(source[key])) {
+                    if (!target[key]) Object.assign(target, { [key]: {} });
+                    this.mergeDeep(target[key], source[key]);
+                } else {
+                    Object.assign(target, { [key]: source[key] });
+                }
+            }
+        }
+  
+        return this.mergeDeep(target, ...sources);
     }
 }