+++ core/src/bake-interfaces.ts
@@ -26,8 +26,10 @@ export interface IBakeEnvironment {
 export interface IIngredientProperties {
     type: string,
     source: BakeVariable,
+    condition?: BakeVariable,
     parameters: Map<string,BakeVariable>,
-    tokens: Map<string,BakeVariable>
+    tokens: Map<string,BakeVariable>,
+    alerts: Map<string,BakeVariable>
 }
 
 export interface IIngredientType {