+++ core/src/index.ts
@@ -7,9 +7,9 @@ import {IngredientManager} from './ingredient-manager'
 import {BaseIngredient} from './base-ingredient'
 import {BaseUtility} from './base-utility'
 import {TagGenerator} from './tag-generator'
-
+import {objToVariableMap} from './utils' 
 
 export {BakeEval, BakeVariable, BaseUtility, DeploymentContext,Logger,
     IBakeAuthentication, IBakeConfig, IBakeEnvironment, IBakePackage, IBakeRegion, IIngredient, IIngredientProperties,
-    IngredientManager, BaseIngredient, TagGenerator
+    IngredientManager, BaseIngredient, TagGenerator, objToVariableMap
 }