+++ ingredient/ingredient-traffic-manager/src/plugin.ts
@@ -5,6 +5,7 @@ import { ARMHelper } from "@azbake/arm-helper";
 import profile from './trf-mgr.json';
 import endpoint from './endpoint.json';
 import { TrafficUtils } from './functions';
+import stockAlerts from "./stockAlerts.json"
 
 export class TrafficManager extends BaseIngredient {
     constructor(name: string, ingredient: IIngredient, ctx: DeploymentContext) {
@@ -40,6 +41,22 @@ export class TrafficManager extends BaseIngredient {
     
             let props = await this._helper.BakeParamsToARMParamsAsync(this._name, this._ctx.Ingredient.properties.parameters);
             props["name"] = {"value": trfutil.get_profile() };
+
+            if (!props["diagnosticsEnabled"])
+                props["diagnosticsEnabled"] = {"value": "yes"}
+
+            if (props["diagnosticsEnabled"].value === "yes") {
+                const ehnUtils = IngredientManager.getIngredientFunction("eventhubnamespace", this._ctx)
+
+                var diagnosticsEventHubNamespace = ehnUtils.get_resource_name("diagnostics");
+                props["diagnosticsEventHubNamespace"] = {"value": diagnosticsEventHubNamespace};
+              
+                var diagnosticsEventHubNamespaceResourceGroup: string
+
+                diagnosticsEventHubNamespaceResourceGroup = await util.resource_group("diagnostics");
+
+                props["diagnosticsEventHubResourceGroup"] = {"value": diagnosticsEventHubNamespaceResourceGroup};                
+            }
 
             await this._helper.DeployTemplate(`${this._name}-profile`, profile, props, await util.resource_group());
 
@@ -81,6 +98,11 @@ export class TrafficManager extends BaseIngredient {
             props["source-type"] = temp["source-type"];
 
             await this._helper.DeployTemplate(`${this._name}-endpoint`, endpoint, props, await util.resource_group());
+
+            let alertTarget = profileName
+            let alertOverrides = this._ingredient.properties.alerts
+            await this._helper.DeployAlerts(this._name, await util.resource_group(), alertTarget, stockAlerts, alertOverrides)
+            
         } catch(error){
             this._logger.error(`deployment failed: ${error}`);
             throw error;